Cambio

+++ SisVentas/src/java/com/ec/controlador/ListaComprasSri.java
@@ -4,7 +4,6 @@
  */
 package com.ec.controlador;
 
-import com.ec.dao.DetalleFacturaDAO;
 import com.ec.dao.xml.Impuesto;
 import com.ec.dao.xml.RetencionXML;
 import com.ec.dao.xml.factura.Detalle;
@@ -13,9 +12,12 @@ import com.ec.dao.xml.factura.TotalImpuesto;
 import com.ec.entidad.CabeceraCompra;
 import com.ec.entidad.ComprasSri;
 import com.ec.entidad.DetalleCompra;
+import com.ec.entidad.DetalleKardex;
+import com.ec.entidad.Kardex;
 import com.ec.entidad.Parametrizar;
 import com.ec.entidad.Producto;
 import com.ec.entidad.Tipoambiente;
+import com.ec.entidad.Tipokardex;
 import com.ec.entidad.docsri.RetencionCompraSri;
 import com.ec.entidad.sri.CabeceraCompraSri;
 import com.ec.entidad.sri.DetalleCompraSri;
@@ -28,16 +30,21 @@ import com.ec.servicio.ServicioCompra;
 import com.ec.servicio.ServicioComprasSri;
 import com.ec.servicio.ServicioDetalleCompra;
 import com.ec.servicio.ServicioDetalleComprasSri;
+import com.ec.servicio.ServicioDetalleKardex;
 import com.ec.servicio.ServicioDetalleRetencionSri;
 import com.ec.servicio.ServicioEstadoFactura;
+import com.ec.servicio.ServicioGeneral;
+import com.ec.servicio.ServicioKardex;
 import com.ec.servicio.ServicioParametrizar;
 import com.ec.servicio.ServicioProducto;
 import com.ec.servicio.ServicioProveedor;
 import com.ec.servicio.ServicioRetencionSri;
 import com.ec.servicio.ServicioTipoAmbiente;
 import com.ec.servicio.ServicioTipoIdentificacionCompra;
+import com.ec.servicio.ServicioTipoKardex;
 import com.ec.untilitario.ArchivoUtils;
 import com.ec.untilitario.AutorizarDocumentos;
+import com.ec.untilitario.TotalKardex;
 import ec.gob.sri.comprobantes.modelo.factura.Factura;
 import ec.gob.sri.comprobantes.ws.aut.Autorizacion;
 import ec.gob.sri.comprobantes.ws.aut.RespuestaComprobante;
@@ -164,6 +171,11 @@ public class ListaComprasSri extends SelectorComposer<Component> {
 
     private static String SRIFACCOMPRAS = "SRIFACCOMPRAS";
     private static String SRIRETENCION = "SRIRETENCION";
+    /*DETALLE DEL KARDEX Y DETALLE KARDEX*/
+    ServicioKardex servicioKardex = new ServicioKardex();
+    ServicioDetalleKardex servicioDetalleKardex = new ServicioDetalleKardex();
+    ServicioTipoKardex servicioTipoKardex = new ServicioTipoKardex();
+    ServicioGeneral servicioGeneral = new ServicioGeneral();
 
     public void doAfterCompose(Component comp) throws Exception {
         super.doAfterCompose(comp);
@@ -237,11 +249,12 @@ public class ListaComprasSri extends SelectorComposer<Component> {
     @Command
     @NotifyChange({"listaComprasSris", "inicio", "fin", "listaComprasSriModel"})
     public void eliminarCabeceraSRI() {
-
-        servicioComprasSri.eliminarCabeceraSri(inicio, fin);
-        findComprasSriByBetweenFecha();
-        Clients.showNotification("Registros eliminados correctamente ",
-                    Clients.NOTIFICATION_TYPE_INFO, null, "middle_center", 1000, true);
+        if (Messagebox.show("Esta seguro de eliminar los registros?", "Question", Messagebox.OK | Messagebox.CANCEL, Messagebox.QUESTION) == Messagebox.OK) {
+            servicioComprasSri.eliminarCabeceraSri(inicio, fin);
+            findComprasSriByBetweenFecha();
+            Clients.showNotification("Registros eliminados correctamente ",
+                        Clients.NOTIFICATION_TYPE_INFO, null, "middle_center", 1000, true);
+        }
     }
 
     @Command
@@ -251,13 +264,46 @@ public class ListaComprasSri extends SelectorComposer<Component> {
             if (servicioCompra.findByNumeroFacturaAndProveedor(item.getCabNumFactura(), item.getCabProveedor()).isEmpty()) {
                 CabeceraCompra cabecera = ArchivoUtils.compraSriToCompra(item);
                 servicioCompra.crear(cabecera);
+
                 for (DetalleCompraSri object : servicioDetalleComprasSri.detallebyCompraSri(item)) {
                     DetalleCompra detalleCompra = ArchivoUtils.detalleSriToDetalleCompra(object, cabecera);
                     servicioDetalleCompra.crear(detalleCompra);
+
+                    /*INGRESAMOS LO MOVIMIENTOS AL KARDEX*/
+                    Kardex kardex = null;
+                    DetalleKardex detalleKardex = null;
+                    Tipokardex tipokardex = servicioTipoKardex.findByTipkSigla("ING");
+                    if (servicioKardex.FindALlKardexs(detalleCompra.getIdProducto()) == null) {
+                        kardex = new Kardex();
+                        kardex.setIdProducto(detalleCompra.getIdProducto());
+                        kardex.setKarDetalle("Inicio de inventario desde la facturacion para el producto: " + detalleCompra.getIdProducto().getProdNombre());
+                        kardex.setKarFecha(new Date());
+                        kardex.setKarFechaKardex(new Date());
+                        kardex.setKarTotal(BigDecimal.ZERO);
+                        servicioKardex.crear(kardex);
+                    }
+                    detalleKardex = new DetalleKardex();
+                    kardex = servicioKardex.FindALlKardexs(detalleCompra.getIdProducto());
+                    detalleKardex.setIdKardex(kardex);
+                    detalleKardex.setDetkFechakardex(item.getCabFecha());
+                    detalleKardex.setDetkFechacreacion(new Date());
+                    detalleKardex.setIdTipokardex(tipokardex);
+                    detalleKardex.setDetkKardexmanual(Boolean.FALSE);
+                    detalleKardex.setDetkDetalles("Aumenta al kardex facturacion con: FACTC-" + item.getCabNumFactura());
+                    detalleKardex.setIdCompra(cabecera);
+
+                    detalleKardex.setDetkCantidad(detalleCompra.getIprodCantidad());
+                    servicioDetalleKardex.crear(detalleKardex);
+                    TotalKardex totales = servicioKardex.totalesForKardex(kardex);
+                    BigDecimal total = totales.getTotalKardex();
+                    kardex.setKarTotal(total);
+                    servicioKardex.modificar(kardex);
                 }
+
             }
 
         }
+        servicioGeneral.corregirProductos();
         Clients.showNotification("Facturas cargadas correctamente ",
                     Clients.NOTIFICATION_TYPE_INFO, null, "middle_center", 1000, true);
 
@@ -2194,6 +2240,18 @@ public class ListaComprasSri extends SelectorComposer<Component> {
 
     public void setRegistrosSeleccionados(Set<CabeceraCompraSri> registrosSeleccionados) {
         this.registrosSeleccionados = registrosSeleccionados;
+    }
+
+    @Command
+    @NotifyChange({"inicio", "fin", "listaComprasSriModel"})
+    public void eliminarCabeceraCabeceraSRI() {
+        if (Messagebox.show("Esta seguro de eliminar los registros?", "Question", Messagebox.OK | Messagebox.CANCEL, Messagebox.QUESTION) == Messagebox.OK) {
+            servicioCabeceraComprasri.eliminarCabeceraComprasSri(inicio, fin);
+            findCabeceraComprasSriByBetweenFecha();
+            Clients.showNotification("Registros eliminados correctamente ",
+                        Clients.NOTIFICATION_TYPE_INFO, null, "middle_center", 1000, true);
+
+        }
     }
 
 }
